The Great Untangling continues; modification of several files believed to relate to biome generation, such as GenLayerBiome and Biome (shocking), have been modified in various ways. Biome now features an expanded enumeration for the different climate zones, WorldType supports a separate type of world generation from the default specifically for the purpose of this mod as well. GenLayer and GenLayerBiomeEdge have also been modified to support the desired climate zone world structure.
Within GenLayerBiome, work is underway constructing the expanded getInts method, which will influence world generation by restricting biomes to specific climate bands dynamically by associating climate zones with horizontal strips of blocks in the world. However, due to ambiguity in the method, a lack of documentation, and the fact I have not dug into the search function yet to find even a single call of the method out of hundreds of classes, have hampered progress. Understanding of the method parameters and the purpose of the output is paramount to fully modifying the method.